Showcase duplicate module compiliation errors
Compiling a module can return a pre-compiled module causing the `Close`
of one CompiledModule to affect the other. This makes it difficult to
implement caching without knowing the module ID that wazero uses
internaly to cache modules. To limit the growth of the cache we want to
close a compiled module to free resources and compile other modules.